perf: extract build logic to separate function
