+++ packages/fx-core/src/component/driver/aad/update.ts
@@ -5,31 +5,19 @@ import { DriverContext } from "../interface/commonArgs";
 import { UpdateAadAppArgs } from "./interface/updateAadAppArgs";
 import { Service } from "typedi";
 import { InvalidParameterUserError } from "./error/invalidParameterUserError";
-import { UpdateAadAppOutput } from "./interface/updateAadAppOutput";
 import { ProgressBarSetting } from "./interface/progressBarSetting";
-import * as fs from "fs-extra";
-import * as path from "path";
 import { AadAppClient } from "./utility/aadAppClient";
 import axios from "axios";
 import { SystemError, UserError, ok, err, FxError, Result } from "@microsoft/teamsfx-api";
 import { UnhandledSystemError, UnhandledUserError } from "./error/unhandledError";
-import { getUuid } from "../../../common/tools";
-import { expandEnvironmentVariable, getEnvironmentVariables } from "../../utils/common";
-import { AadManifestHelper } from "../../resource/aadApp/utils/aadManifestHelper";
-import { AADManifest } from "../../resource/aadApp/interfaces/AADManifest";
-import { MissingFieldInManifestUserError } from "./error/invalidFieldInManifestError";
-import isUUID from "validator/lib/isUUID";
 import { hooks } from "@feathersjs/hooks/lib";
 import { addStartAndEndTelemetry } from "../middleware/addStartAndEndTelemetry";
 import { getLocalizedString } from "../../../common/localizeUtils";
 import { logMessageKeys, descriptionMessageKeys } from "./utility/constants";
-import { MissingEnvUserError } from "./error/missingEnvError";
+import { buildAadManifest, loadCurrentState } from "./utility/buildAadManifest";
 
 const actionName = "aadApp/update"; // DO NOT MODIFY the name
 const helpLink = "https://aka.ms/teamsfx-actions/aadapp-update";
-const driverConstants = {
-  generateManifestFailedMessageKey: "driver.aadApp.error.generateManifestFailed",
-};
 
 // logic from src\component\resource\aadApp\aadAppManifestManager.ts
 @Service(actionName) // DO NOT MODIFY the service name
@@ -57,41 +45,16 @@ export class UpdateAadAppDriver implements StepDriver {
       await progressHandler?.start();
       await progressHandler?.next(progressBarSettings.stepMessages.shift());
       context.logProvider?.info(getLocalizedString(logMessageKeys.startExecuteDriver, actionName));
+      const state = loadCurrentState();
 
       this.validateArgs(args);
       const aadAppClient = new AadAppClient(context.m365TokenProvider);
-      const state = this.loadCurrentState();
 
-      const manifestAbsolutePath = this.getAbsolutePath(
+      const manifest = await buildAadManifest(
+        context,
         args.manifestTemplatePath,
-        context.projectPath
+        args.outputFilePath
       );
-      const manifest = await this.loadManifest(manifestAbsolutePath, state);
-      const warningMessage = AadManifestHelper.validateManifest(manifest);
-      if (warningMessage) {
-        warningMessage.split("\n").forEach((warning) => {
-          context.logProvider?.warning(warning);
-        });
-      }
-
-      if (!manifest.id || !isUUID(manifest.id)) {
-        throw new MissingFieldInManifestUserError(actionName, "id", helpLink);
-      }
-
-      // Output actual manifest to project folder first for better troubleshooting experience
-      const outputFileAbsolutePath = this.getAbsolutePath(args.outputFilePath, context.projectPath);
-      await fs.ensureDir(path.dirname(outputFileAbsolutePath));
-      await fs.writeFile(outputFileAbsolutePath, JSON.stringify(manifest, null, 4), "utf8");
-      context.logProvider?.info(
-        getLocalizedString(logMessageKeys.outputAadAppManifest, outputFileAbsolutePath)
-      );
-
-      if (args.onlyBuild) {
-        return {
-          result: ok(new Map()),
-          summaries: [],
-        };
-      }
 
       // MS Graph API does not allow adding new OAuth permissions and pre authorize it within one request
       // So split update AAD app to two requests:
@@ -181,51 +144,6 @@ export class UpdateAadAppDriver implements StepDriver {
     }
   }
 
-  private loadCurrentState(): UpdateAadAppOutput {
-    return {
-      AAD_APP_ACCESS_AS_USER_PERMISSION_ID: process.env.AAD_APP_ACCESS_AS_USER_PERMISSION_ID,
-    };
-  }
-
-  private async loadManifest(
-    manifestPath: string,
-    state: UpdateAadAppOutput
-  ): Promise<AADManifest> {
-    let generatedNewPermissionId = false;
-    try {
-      const manifestTemplate = await fs.readFile(manifestPath, "utf8");
-      const permissionIdPlaceholderRegex = /\${{ *AAD_APP_ACCESS_AS_USER_PERMISSION_ID *}}/;
-
-      // generate a new permission id if there's no one in env and manifest needs it
-      if (!process.env.AAD_APP_ACCESS_AS_USER_PERMISSION_ID) {
-        const matches = permissionIdPlaceholderRegex.exec(manifestTemplate);
-        if (matches) {
-          const permissionId = getUuid();
-          process.env.AAD_APP_ACCESS_AS_USER_PERMISSION_ID = permissionId;
-          state.AAD_APP_ACCESS_AS_USER_PERMISSION_ID = permissionId;
-          generatedNewPermissionId = true;
-        }
-      }
-
-      const manifestString = expandEnvironmentVariable(manifestTemplate);
-      this.validateManifestString(manifestString);
-      const manifest: AADManifest = JSON.parse(manifestString);
-      AadManifestHelper.processRequiredResourceAccessInManifest(manifest);
-      return manifest;
-    } finally {
-      if (generatedNewPermissionId) {
-        // restore environment variable to avoid impact to other code
-        delete process.env.AAD_APP_ACCESS_AS_USER_PERMISSION_ID;
-      }
-    }
-  }
-
-  private getAbsolutePath(relativeOrAbsolutePath: string, projectPath: string) {
-    return path.isAbsolute(relativeOrAbsolutePath)
-      ? relativeOrAbsolutePath
-      : path.join(projectPath, relativeOrAbsolutePath);
-  }
-
   private getProgressBarSetting(): ProgressBarSetting {
     return {
       title: getLocalizedString("driver.aadApp.progressBar.updateAadAppTitle"),
@@ -233,17 +151,5 @@ export class UpdateAadAppDriver implements StepDriver {
         getLocalizedString("driver.aadApp.progressBar.updateAadAppStepMessage"), // step 1
       ],
     };
-  }
-
-  private validateManifestString(manifestString: string) {
-    const unresolvedEnvironmentVariable = getEnvironmentVariables(manifestString);
-    if (unresolvedEnvironmentVariable && unresolvedEnvironmentVariable.length > 0) {
-      throw new MissingEnvUserError(
-        actionName,
-        unresolvedEnvironmentVariable,
-        helpLink,
-        driverConstants.generateManifestFailedMessageKey
-      );
-    }
   }
 }
