+++ packages/fx-core/src/core/FxCoreImplementV3.ts
@@ -73,6 +73,7 @@ import { QuestionMW } from "../component/middleware/questionMW";
 import { getQuestionsForCreateProjectV2 } from "./middleware/questionModel";
 import { getQuestionsForInit, getQuestionsForProvisionV3 } from "../component/question";
 import { isFromDevPortalInVSC } from "../component/developerPortalScaffoldUtils";
+import { buildAadManifest } from "../component/driver/aad/utility/buildAadManifest";
 
 export class FxCoreV3Implement {
   tools: Tools;
@@ -210,43 +211,6 @@ export class FxCoreV3Implement {
       ctx!.envVars = inputs.envVars;
       return err(res.error);
     }
-  }
-
-  @hooks([ErrorHandlerMW, ProjectConsolidateMW, EnvLoaderMW(false)])
-  async previewAadManifest(inputs: Inputs): Promise<Result<Void, FxError>> {
-    setCurrentStage(Stage.deployAad);
-    inputs.stage = Stage.deployAad;
-    const updateAadClient = Container.get<UpdateAadAppDriver>("aadApp/update");
-    // In V3, the aad.template.json exist at .fx folder, and output to root build folder.
-    const manifestTemplatePath: string = inputs.AAD_MANIFEST_FILE
-      ? inputs.AAD_MANIFEST_FILE
-      : path.join(inputs.projectPath!, AadConstants.DefaultTemplateFileName);
-    if (!(await fs.pathExists(manifestTemplatePath))) {
-      return err(new NoAadManifestExistError(manifestTemplatePath));
-    }
-    await fs.ensureDir(path.join(inputs.projectPath!, "build"));
-    const manifestOutputPath: string = path.join(
-      inputs.projectPath!,
-      "build",
-      `aad.${inputs.env}.json`
-    );
-    const inputArgs: UpdateAadAppArgs = {
-      manifestTemplatePath: manifestTemplatePath,
-      outputFilePath: manifestOutputPath,
-      onlyBuild: true,
-    };
-    const contextV3: DriverContext = {
-      azureAccountProvider: TOOLS.tokenProvider.azureAccountProvider,
-      m365TokenProvider: TOOLS.tokenProvider.m365TokenProvider,
-      ui: TOOLS.ui,
-      logProvider: TOOLS.logProvider,
-      telemetryReporter: TOOLS.telemetryReporter!,
-      projectPath: inputs.projectPath as string,
-      platform: Platform.VSCode,
-    };
-    const res = await updateAadClient.run(inputArgs, contextV3);
-    if (res.isErr()) return err(res.error);
-    return ok(Void);
   }
 
   @hooks([
@@ -471,6 +435,32 @@ export class FxCoreV3Implement {
       });
 
     writeStream.end();
+    return ok(Void);
+  }
+
+  async previewAadManifest(inputs: Inputs): Promise<Result<Void, FxError>> {
+    const manifestTemplatePath: string = inputs.AAD_MANIFEST_FILE
+      ? inputs.AAD_MANIFEST_FILE
+      : path.join(inputs.projectPath!, AadConstants.DefaultTemplateFileName);
+    if (!(await fs.pathExists(manifestTemplatePath))) {
+      return err(new NoAadManifestExistError(manifestTemplatePath));
+    }
+    await fs.ensureDir(path.join(inputs.projectPath!, "build"));
+    const manifestOutputPath: string = path.join(
+      inputs.projectPath!,
+      "build",
+      `aad.${inputs.env}.json`
+    );
+    const contextV3: DriverContext = {
+      azureAccountProvider: TOOLS.tokenProvider.azureAccountProvider,
+      m365TokenProvider: TOOLS.tokenProvider.m365TokenProvider,
+      ui: TOOLS.ui,
+      logProvider: TOOLS.logProvider,
+      telemetryReporter: TOOLS.telemetryReporter!,
+      projectPath: inputs.projectPath as string,
+      platform: Platform.VSCode,
+    };
+    await buildAadManifest(contextV3, manifestTemplatePath, manifestOutputPath);
     return ok(Void);
   }
 }
